Add files via upload
This is a very simple notebook where I've defined the function newtonian_potential(x, Epsilon) which takes as input Epsilon as a 2-d matrix, computed on a square grid, whose side is x. Then it computes the 2-d Newtonian potential by simply numerical integrating. I also plot the comparison between the exact solution  and the point-mass limit, as a benchmark. Next step: obtain the shear, that is basically the gradient of the 2-d potential times a normalization constant which  contains angular distances between the source and the observer, the source and the lens, and the lens and the observer. They can be easily computed say, from CLASS